ISSUE #54. Base plugin has been added
