+++ src/docs/asciidoc/parts/base-plugin.adoc
@@ -1,0 +1,33 @@
+== Base plugin
+
+If the convention of adding a `gradleTest` with it's default settings are not suitable for your
+environment, then the base plugin can be considered. This will only add the
+`additionalGradleTestSet` method.
+The GradleTest Base Plugin is available in the https://plugins.gradle.org/plugin/org.ysb33r.gradletest.base[Gradle Plugins] repository.
+To use it, add the following snippet into your build script.
+
+[source,groovy,subs="attributes"]
+----
+plugins {
+  id 'org.ysb33r.gradletest.base' version '{revnumber}'
+}
+----
+
+When your setup is more complex and the plugins block does not work OR if you are using Gradle 2.0 use the following instead
+
+[source,groovy,subs="attributes"]
+----
+buildscript {
+  repositories {
+    maven {
+      url "https://plugins.gradle.org/m2/"
+    }
+  }
+
+  dependencies {
+    classpath 'org.ysb33r.gradle:gradletest:{revnumber}'
+  }
+}
+
+apply plugin: 'org.ysb33r.gradletest.base'
+----
