+++ src/main/resources/META-INF/gradle-plugins/org.ysb33r.gradletest.base.properties
@@ -1,0 +1,1 @@
+implementation-class=org.ysb33r.gradle.gradletest.GradleTestBasePlugin
